+++ src/de/usd/cstchef/operations/setter/HttpGetSetter.java
@@ -10,7 +10,7 @@ import de.usd.cstchef.operations.Operation.OperationInfos;
 import de.usd.cstchef.operations.OperationCategory;
 
 @OperationInfos(name = "HTTP GET Param", category = OperationCategory.SETTER, description = "Sets a GET parameter to the specified value.")
-public class GetSetter extends SetterOperation {
+public class HttpGetSetter extends SetterOperation {
 
 	private JCheckBox addIfNotPresent;
 	private JCheckBox urlEncode;